[REF] core: extract discover_local_tool_providers to core.mcp_gateway
